Makefile + destroyZombieHorde in exercice 01

+++ ex01/zombieHorde.cpp
@@ -17,7 +17,7 @@ Zombie	*zombieHorde(int N, std::string name)
 	Zombie	*horde;
 
 	horde = new Zombie[N];
-	for (size_t i = 0; i < N; ++i)
+	for (int i = 0; i < N; ++i)
 	{
 		horde[i].setName(name);
 	}
